requirements/dev.txt
-e git+ssh://git@github.com/mitsuhiko/flask@577581665ce745898c29ca502e0a3aba1f16677e#egg=Flask-dev
Flask-Admin==1.0.8
Flask-Alembic==1.0.1
Flask-Bootstrap==3.1.1.2
Flask-Login==0.2.11
Flask-Migrate==1.2.0
Flask-Moment==0.3.1
Flask-OpenID==1.2.1
Flask-Rauth==0.3.2
Flask-SQLAlchemy==1.0
Flask-Script==2.0.5
Flask-Testing==0.4.1
Flask-WTF==0.9.5
ForgeryPy==0.1
Jinja2==2.7.3
Mako==0.9.1
MarkupSafe==0.23
Pygments==1.6
SQLAlchemy==0.9.6
WTForms==1.0.5
Werkzeug==0.9.6
alembic==0.6.5
astroid==1.1.1
blinker==1.3
bouncer==0.1.10
click==2.4
coverage==3.7.1
flask-bouncer==0.1.13
gunicorn==19.0.0
httpie==0.8.0
itsdangerous==0.24
logilab-common==0.61.0
nose==1.3.3
psycopg2==2.5.3
pylint==1.2.1
python-dateutil==2.2
python3-openid==3.0.4
rauth==0.7.0
requests==2.3.0
simplejson==3.6.0
six==1.7.3
